+++ test/post-deploy.test.mjs
@@ -79,6 +79,13 @@ describe('Helix RUM Collector Post-Deploy Tests', () => {
     expect(response).to.have.status(201);
   });
 
+  it('CORS headers are set', async () => {
+    const response = await request(`https://${domain}`)
+      .options('/');
+    expect(response).to.have.status(200);
+    expect(response).to.have.header('access-control-allow-origin', '*');
+  });
+
   it('robots.txt denies everything', async () => {
     const response = await request(`https://${domain}`)
       .get('/robots.txt');