+++ test/s3-logger.test.mjs
@@ -50,7 +50,7 @@ describe('Test S3 Logger', () => {
     assert.equal(logged.target.toString(), 'http://www.foo.com/target');
     assert.equal(logged.source.toString(), 'http://www.foo.com/source');
     assert.equal(logged.id, 'someid');
-    assert.equal(logged.user_agent, 'desktop');
+    assert.equal(logged.user_agent, 'desktop:windows');
     assert.deepEqual(logged.foo, ['b', 'ar']);
   });
 });